requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
comm==0.2.3
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
faiss-cpu==1.7.4
fastapi==0.120.4
ffmpy==0.6.4
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gradio==4.8.0
gradio_client==0.7.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.31.0
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langsmith==0.0.87
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.10.1
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.2
numpy==1.26.4
orjson==3.11.4
packaging==23.2
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pypdf==3.17.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==5.1.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.49.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tomlkit==0.12.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==11.0.3
yarl==1.22.0
zipp==3.23.0